+++ src/ossl/ecdsa.rs
@@ -283,7 +283,6 @@ impl EcdsaOperation {
 
     /// Creates a new `EcdsaOperation` for verification with a pre-supplied
     /// signature.
-    #[cfg(feature = "pkcs11_3_2")]
     pub fn verify_signature_new(
         mech: &CK_MECHANISM,
         key: &Object,
@@ -491,7 +490,6 @@ impl Verify for EcdsaOperation {
     }
 }
 
-#[cfg(feature = "pkcs11_3_2")]
 impl VerifySignature for EcdsaOperation {
     fn verify(&mut self, data: &[u8]) -> Result<()> {
         self.verify_internal(data, None)