+++ src/tests/rsa.rs
@@ -214,13 +214,9 @@ fn test_rsa_operations() {
     let ret = sig_verify(session, pub_key_handle, &msg, &sig, &mechanism);
     assert_eq!(ret, CKR_OK);
 
-    #[cfg(feature = "pkcs11_3_2")]
-    {
-        /* Re-Verify using the SignatureVerification APIs */
-        let ret =
-            sig_verifysig(session, pub_key_handle, &msg, &sig, &mechanism);
-        assert_eq!(ret, CKR_OK);
-    }
+    /* Re-Verify using the SignatureVerification APIs */
+    let ret = sig_verifysig(session, pub_key_handle, &msg, &sig, &mechanism);
+    assert_eq!(ret, CKR_OK);
 
     let signed =
         ret_or_panic!(sig_gen(session, pri_key_handle, &msg, &mechanism));